Now validates numeric input and fixed a bug with future start dates
Now checks to see if total words required and words written so far are
entered as numbers.  Also fixed a bug where if you entered a future
project start date, it would throw an error when telling you how many
days you had left at the end of the script.